scripts/requirements-ingest.txt
requests==2.32.3
beautifulsoup4==4.12.3
google-generativeai==0.8.3
boto3==1.34.131
tqdm==4.66.4
tenacity==8.5.0
